split out tips
